pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-supervisor"
version = "0.1.0"
description = "Two-agent Codex supervisor with protocol automation and logging"
authors = [{name = "5310S"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
"codex-supervisor" = "supervisor:main"
"codex-supervisor-gui" = "gui:main"
